feat(issue139): Add working charts still need tooltips and genres
