+++ src/components/Log.tsx
@@ -7,7 +7,7 @@ import { LogReducer } from "../config/logStore";
 import { useMDDispatch, useMDState } from "../config/store";
 import type { DiaryAdd, MediaSelected } from "../config/types";
 import useFuegoUser from "../hooks/useFuegoUser";
-import { fuegoDiaryAdd } from "../interfaces/fuegoActions";
+import { fuegoDiaryAdd } from "../interfaces/fuegoMDActions";
 import { fetcher, spotifyFetch } from "../utils/fetchers";
 import Info from "./Info";
 import LogFields from "./LogFields";
